Moved to new solution with API Login Server

+++ Services/Application/GameServer/GameServerService.cs
@@ -7,8 +7,17 @@ using ServerYourWorldMMORPG.Services.Application.Interfaces;
 using ServerYourWorldMMORPG.Utils;
 using System.Net;
 using System.Net.Sockets;
-using System.Numerics;
 using System.Text;
+
+public enum ServerToClient : ushort
+{
+	playerSpawned = 1,
+}
+
+public enum ClientToServerId : ushort
+{
+	name = 1,
+}
 
 namespace ServerYourWorldMMORPG.Services.Application.GameServer
 {
@@ -36,11 +45,13 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			name = _gameServerSettings.Name;
 			isRunning = false;
 
-			_tcpListener = new TcpListener(IPAddress.Parse(_gameServerSettings.IpAddress), _gameServerSettings.Port);
+			_tcpListener = new TcpListener(IPAddress.Parse(_gameServerSettings.IpAddress), _loginServerSettings.Port);
 			IPEndPoint udpEndPoit = new IPEndPoint(IPAddress.Parse(_gameServerSettings.IpAddress), _gameServerSettings.Port);
 			_udpListener = new UdpClient(udpEndPoit);
 			_connectedUserSessions = new Dictionary<string, UserSession>();
-			StartListeningForUdp();
+
+			//RiptideLogger.Initialize(Console.WriteLine, false);
+			//server = new Server();
 		}
 
 		public bool ServerStatus()
@@ -51,23 +62,32 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 		public async Task StartServer()
 		{
 			isRunning = true;
+			//server.Start((ushort)_gameServerSettings.Port, (ushort)_gameServerSettings.MaxPlayers);
+			//server.ClientDisconnected += PlayerLeft;
+			//server.ClientConnected += PlayerConnected;
+
 			_tcpListener.Start();
-			_udpListener.BeginReceive(OnUdpDataReceived, null);
+			StartListeningForUdp();
 			ConsoleUtility.Print($"{name} started.");
 
-			// Register with the central login server
-			//await RegisterWithLoginServer();
+			////Register with the central login server
+			////await RegisterWithLoginServer();
+			////_tcpListener.BeginAcceptTcpClient();
 
 			while (isRunning)
 			{
+				CheckClientConnections();
 				var client = await _tcpListener.AcceptTcpClientAsync();
+				//Task.WhenAll([CheckClientConnections(), OnClientConnect(client)]);
 				Task.Run(() => OnClientConnect(client), _cancellationTokenSource.Token);
 			}
 		}
-
 		public async Task StopServer()
 		{
 			isRunning = false;
+			//server.Stop();
+			_udpListener.Close();
+			_udpListener?.Dispose();
 			_tcpListener.Stop();
 
 			foreach (var userSession in _connectedUserSessions.Values)
@@ -85,7 +105,7 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			{
 				foreach (var client in _connectedUserSessions)
 				{
-					ConsoleUtility.Print($"Connected client: {client.Key} UDP Endpoint: {client.Value.UdpEndPoint}");
+					ConsoleUtility.Print($"Connected client: {client.Key} UDP Endpoint: {client.Value.ClientIpEndPoint}");
 					if (client.Value.PlayerNetworkObject.NetworkObjectId != null)
 					{
 						ConsoleUtility.Print($"Network Object ID: {client.Value.PlayerNetworkObject.NetworkObjectId}");
@@ -96,6 +116,7 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			{
 				ConsoleUtility.Print("There are no connected clients yet!");
 			}
+
 			//var clients = new List<UserClient>();
 			//foreach (var kvp in _connectedUserSessions)
 			//{
@@ -111,85 +132,12 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			//return clients;
 		}
 
-		public async Task SendMessage(string clientId, string message, ProtocolType protocol = ProtocolType.Udp)
+		public int NumberOfConnectedClients()
 		{
-			if (protocol == ProtocolType.Tcp)
-			{
-				var client = GetTcpClientFromId(clientId);
-				if (client == null || !client.Connected) return;
-
-				NetworkStream stream = client.GetStream();
-				if (stream.CanWrite)
-				{
-					byte[] buffer = Encoding.UTF8.GetBytes(message);
-					await stream.WriteAsync(buffer, 0, buffer.Length);
-					ConsoleUtility.Print($"{protocol}/TCP message: {message}");
-				}
-			}
-			else if (protocol == ProtocolType.Udp)
-			{
-				if (_connectedUserSessions.TryGetValue(clientId, out var userSession))
-				{
-					byte[] data = Encoding.UTF8.GetBytes(message);
-					await _udpListener.SendAsync(data, data.Length, userSession.UdpEndPoint);
-					ConsoleUtility.Print($"{protocol}/UDP message: {message}");
-				}
-				else
-				{
-					ConsoleUtility.Print($"UDP endpoint not found for client: {clientId}");
-				}
-			}
+			return _connectedUserSessions.Count();
 		}
 
-		public async Task BroadcastMessage(string message, string exceptUser = null)
-		{
-			foreach (var client in _connectedUserSessions)
-			{
-				if (exceptUser == client.Key) continue;
-				await SendMessage(client.Key, message);
-			}
-		}
-
-		private void StartListeningForUdp()
-		{
-			_udpListener.BeginReceive(OnUdpDataReceived, null);
-		}
-
-		private void OnUdpDataReceived(IAsyncResult result)
-		{
-			IPEndPoint clientEndPoint = new IPEndPoint(IPAddress.Any, 0);
-			byte[] data = new byte[0];
-			try
-			{
-				data = _udpListener.EndReceive(result, ref clientEndPoint);
-			}
-			catch (Exception ex)
-			{
-				ConsoleUtility.Print($"Host disconected: {ex}");
-				//_connectedUserSessions.TryGetValue(clientEndPoint, out UserSession value);
-				//RemovePlayer();
-			}
-
-			// Process the received data
-			string message = Encoding.UTF8.GetString(data);
-			ConsoleUtility.Print($"Received UDP message from {clientEndPoint}: {message}");
-
-			if (message == "RequestSessionId")
-			{
-				ClientRequsestedSessionId(clientEndPoint);
-			}
-
-			var clientEntry = _connectedUserSessions.FirstOrDefault(entry => entry.Value.UdpEndPoint.Equals(clientEndPoint));
-
-			if (clientEntry.Key != null && message != "RequestSessionId")
-			{
-				string sessionId = clientEntry.Key;
-				ProcessServerMessage(sessionId, message);
-			}
-
-			_udpListener.BeginReceive(OnUdpDataReceived, null);
-		}
-
+		// TCP Methods START
 		private TcpClient? GetTcpClientFromId(string clientId)
 		{
 			if (_connectedUserSessions.TryGetValue(clientId, out UserSession? userSession))
@@ -219,7 +167,7 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 					{
 						string message = Encoding.UTF8.GetString(buffer, 0, bytesRead);
 						//ConsoleUtility.Print($"Message: {message} From: {clientId}");
-						ProcessServerMessage(sessionId, message);
+						//ProcessServerMessage(sessionId, message);
 					}
 				}
 			}
@@ -233,34 +181,207 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			}
 		}
 
-		private void ProcessServerMessage(string sessionId, string message)
+		private void CleanUpClientResources(string sessionId, NetworkStream stream)
 		{
+			if (_connectedUserSessions.TryGetValue(sessionId, value: out UserSession? userSessionToRemove))
+			{
+				_networkObjectService.RemoveNetworkObject(userSessionToRemove.PlayerNetworkObject);
+			}
+
+			if (userSessionToRemove == null) return;
+
+			stream.Close();
+			userSessionToRemove.TcpClient.Close();
+			_connectedUserSessions.Remove(sessionId);
+			ConsoleUtility.Print($"Client {sessionId} has disconnected!");
+		}
+		// TCP Methods END
+
+		public async Task SendMessage(string sessionId, string message, ProtocolType protocol = ProtocolType.Udp)
+		{
+			if (protocol == ProtocolType.Tcp)
+			{
+				var client = GetTcpClientFromId(sessionId);
+				if (client == null || !client.Connected) return;
+
+				NetworkStream stream = client.GetStream();
+				if (stream.CanWrite)
+				{
+					byte[] buffer = Encoding.UTF8.GetBytes(message);
+					await stream.WriteAsync(buffer, 0, buffer.Length);
+					ConsoleUtility.Print($"{protocol}/TCP message: {message}");
+				}
+			}
+			else if (protocol == ProtocolType.Udp)
+			{
+				if (_connectedUserSessions.TryGetValue(sessionId, out UserSession userSession))
+				{
+					byte[] data = Encoding.UTF8.GetBytes(message);
+					_udpListener.SendAsync(data, data.Length, userSession.ClientIpEndPoint);
+					if (!message.Contains("UpdatePosition"))
+					{
+						ConsoleUtility.Print($"{protocol}/UDP message: {message}");
+					}
+				}
+				else
+				{
+					ConsoleUtility.Print($"UDP endpoint not found for client: {sessionId}");
+				}
+			}
+		}
+
+		public async Task BroadcastMessage(string message, string exceptUser = null)
+		{
+			foreach (var client in _connectedUserSessions)
+			{
+				if (exceptUser == client.Key) continue;
+				await SendMessage(client.Key, message);
+			}
+		}
+
+		private void StartListeningForUdp()
+		{
+			_udpListener.BeginReceive(OnUdpDataReceived, null);
+		}
+
+		private async void OnUdpDataReceived(IAsyncResult result)
+		{
+			//IPEndPoint clientEndPoint = new IPEndPoint(IPAddress.Any, 0);
+
+			//if (_udpListener != null)
+			//{
+			//	byte[] data = _udpListener.EndReceive(result, ref clientEndPoint);
+			//	string message = Encoding.UTF8.GetString(data);
+			//	ProcessServerMessage(clientEndPoint, message);
+			//	_udpListener.BeginReceive(OnUdpDataReceived, null);
+			//}
+			//else
+			//{
+			//	DisconnectClientByIPEndPoint(clientEndPoint);
+			//}
+
+			IPEndPoint clientEndPoint = new IPEndPoint(IPAddress.Any, 0);
+			try
+			{
+				byte[] data = _udpListener.EndReceive(result, ref clientEndPoint);
+				string message = Encoding.UTF8.GetString(data);
+				ProcessServerMessage(clientEndPoint, message);
+			}
+			catch (SocketException ex)
+			{
+				ConsoleUtility.Print($"SocketException in OnUdpDataReceived: {ex.Message}");
+				DisconnectClientByIPEndPoint(clientEndPoint);
+			}
+			catch (Exception ex)
+			{
+				ConsoleUtility.Print($"Exception in OnUdpDataReceived: {ex.Message}");
+				DisconnectClientByIPEndPoint(clientEndPoint);
+			}
+			finally
+			{
+				if (_udpListener != null)
+				{
+					_udpListener.BeginReceive(OnUdpDataReceived, null);
+				}
+			}
+		}
+
+		private async Task ProcessServerMessage(IPEndPoint clientEndPoint, string message)
+		{
+			if (!message.Contains("UpdatePosition"))
+			{
+				ConsoleUtility.Print($"Received UDP message from {clientEndPoint}: {message}");
+			}
+
+			if (message == "RequestSessionId" && !IsEndPointDuplicate(clientEndPoint))
+			{
+				await ClientRequsestedSessionId(clientEndPoint);
+			}
+
 			string[] splitMessage = message.Split('|');
-			string command = splitMessage[0];
-			string data = splitMessage[1];
+			string sessionId = splitMessage[0];
+
+			if (IsSessionIdValid(sessionId))
+			{
+				if (_connectedUserSessions.TryGetValue(sessionId, value: out UserSession userSession))
+				{
+					if (userSession.ClientIpEndPoint != clientEndPoint)
+					{
+						userSession.LastMessageReceived = DateTime.Now;
+
+						ProcessServerMessageForLoggedInUser(sessionId, splitMessage);
+					}
+					else
+					{
+						ConsoleUtility.Print($"!!!!!! Hacker detected with IP: {clientEndPoint.Address} Port: {clientEndPoint.Port}");
+					}
+				}
+			}
+		}
+
+		private bool IsEndPointDuplicate(IPEndPoint clientEndPoint)
+		{
+			bool endPointExistance = _connectedUserSessions.Values
+				.Where(session => session.ClientIpEndPoint == clientEndPoint)
+				.Count() > 1;
+
+			return endPointExistance;
+		}
+
+		private async Task ProcessServerMessageForLoggedInUser(string sessionId, string[] splitMessage)
+		{
+			string command = splitMessage[1];
 
 			switch (command)
 			{
 				case "UpdatePosition":
-					UpdatePosition(data);
+					UpdatePosition(splitMessage[2]);
 					break;
-				case "NewPlayer":
-					string newPlayerId = splitMessage[1];
-					Vector3 spawnPos = new Vector3(float.Parse(splitMessage[2]), float.Parse(splitMessage[3]), float.Parse(splitMessage[4]));
-					//NotifyClientToSpawnPlayer(newPlayerId);
-					break;
-				//case command.StartsWith('PlayerConnected'):
 				case "PlayerConnected":
-					PlayerSpawnRequest(data);
+					PlayerSpawnRequest(sessionId);
 					break;
 				case "PlayerDisconnected":
-					RemovePlayer(sessionId);
+					PlayerDisconnect(sessionId);
+					break;
+				case "HeartBeat":
+					HeartBeatPacket(sessionId);
+					break;
+				case "PlayersConnected":
+					string message = $"PlayersConnected|{NumberOfConnectedClients()}";
+					SendMessage(sessionId, message);
 					break;
 				default:
-					string data2 = splitMessage[1];
-					ConsoleUtility.Print($"Unknown packet received: {data2} From: {sessionId}");
+					ConsoleUtility.Print($"Unknown packet received: {splitMessage} From: {sessionId}");
 					break;
 			}
+		}
+
+		public void CheckClientConnections()
+		{
+			foreach (var session in _connectedUserSessions)
+			{
+				if (DateTime.UtcNow - session.Value.LastMessageReceived > TimeSpan.FromSeconds(30))
+				{
+					ConsoleUtility.Print($"Missed heartbeat disconnecting client: {session.Key}");
+					PlayerDisconnect(session.Key);
+				}
+			}
+		}
+
+		private bool IsSessionIdValid(string sessionId)
+		{
+			if (_connectedUserSessions.TryGetValue(sessionId, value: out UserSession? userSession))
+			{
+				return true;
+			}
+
+			return false;
+		}
+
+		private async Task HeartBeatPacket(string sessionId)
+		{
+			CheckClientConnections();
+			await SendMessage(sessionId, "HeartBeatReceived");
 		}
 
 		private async Task UpdatePosition(string message)
@@ -284,30 +405,49 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			}
 		}
 
-		private async void RemovePlayer(string sessionId)
+		private async void PlayerDisconnect(string sessionId)
 		{
-			_connectedUserSessions.TryGetValue(sessionId, value: out var userSession);
-
-			string message = $"ClientDisconnect|{userSession.PlayerNetworkObject.NetworkObjectId}";
-			await BroadcastMessage(message, sessionId);
-
-			_networkObjectService.RemoveNetworkObject(userSession.PlayerNetworkObject);
-			_connectedUserSessions.Remove(sessionId);
+			if (_connectedUserSessions.TryGetValue(sessionId, value: out UserSession userSession))
+			{
+				if (userSession.PlayerNetworkObject != null)
+				{
+					string message = $"ClientDisconnect|{userSession.PlayerNetworkObject.NetworkObjectId}";
+					await BroadcastMessage(message, sessionId);
+					_networkObjectService.RemoveNetworkObject(userSession.PlayerNetworkObject);
+					_connectedUserSessions.Remove(sessionId);
+				}
+			}
 		}
 
-		private void CleanUpClientResources(string sessionId, NetworkStream stream)
+		public async Task DisconnectClientByIPEndPoint(IPEndPoint clientEndPoint)
 		{
-			if (_connectedUserSessions.TryGetValue(sessionId, value: out UserSession? userSessionToRemove))
+			var sessionToDisconnect = _connectedUserSessions.Values
+				.FirstOrDefault(session => session.ClientIpEndPoint.Equals(clientEndPoint));
+
+			if (sessionToDisconnect != null)
 			{
-				_networkObjectService.RemoveNetworkObject(userSessionToRemove.PlayerNetworkObject);
+				sessionToDisconnect.TcpClient?.Close();
+
+				string sessionId = GetSessionIdByEndPoint(clientEndPoint);
+				if (!string.IsNullOrEmpty(sessionId))
+				{
+					string message = $"ClientDisconnect|{sessionToDisconnect.PlayerNetworkObject.NetworkObjectId}";
+					await BroadcastMessage(message, sessionId);
+					_connectedUserSessions.Remove(sessionId);
+				}
 			}
+		}
 
-			if (userSessionToRemove == null) return;
-
-			stream.Close();
-			userSessionToRemove.TcpClient.Close();
-			_connectedUserSessions.Remove(sessionId);
-			ConsoleUtility.Print($"Client {sessionId} has disconnected!");
+		public string GetSessionIdByEndPoint(IPEndPoint clientEndPoint)
+		{
+			foreach (var session in _connectedUserSessions)
+			{
+				if (session.Value.ClientIpEndPoint.Equals(clientEndPoint))
+				{
+					return session.Key;  // Return the session ID (key)
+				}
+			}
+			return null;  // Return null if no matching session is found
 		}
 
 		private string GenerateClientId()
@@ -338,7 +478,7 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			}
 			else
 			{
-				userSession = new UserSession { UdpEndPoint = clientEndPoint };
+				userSession = new UserSession { ClientIpEndPoint = clientEndPoint };
 				protocol = ProtocolType.Udp;
 			}
 
@@ -353,6 +493,7 @@ namespace ServerYourWorldMMORPG.Services.Application.GameServer
 			await ProvideSessionIdToClient(sessionId, false, null, clientEndPoint);
 		}
 
+		// Currently not in use
 		private async Task RegisterWithLoginServer()
 		{
 			using var tcpClient = new TcpClient();
